requirements.txt
twilio>=9.2.0
livekit>=0.16.3
livekit-agents>=0.9.0
livekit-plugins-openai>=0.7.0
fastapi>=0.109.0
uvicorn>=0.27.0
python-multipart>=0.0.6
python-dotenv>=1.0.0
aiohttp>=3.10.0
numpy>=1.24.0
audioop-lts>=0.2.1
websockets>=12.0
colorlog>=6.8.0

